app/requirements.txt
tensorflow==2.17.0
pyspark==3.5.1
kafka-python==2.0.2
opencv-python
numpy
matplotlib
pillow
scikit-learn
